contract/budget: Add link to create prefilled budget with contract data
This feature provides the possibility to create a one by one contract:budget easily instead of manually create contract and budget.
The contract needs to be saved to transfer the information to the budget form. 
The form gets the fields title, budget total and contract filled automatically.

Furthermore the budget total can't extend the contract total anymore. When this case occurs the budget can't be saved.